integrations/driver/requirements.txt
requests==2.28.1
pydantic==1.10.2
PyYAML==6.0
typing-extensions==4.5.0
pytz==2023.3.post1
